requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic
pydantic-settings
requests
beautifulsoup4
pytrends
python-dotenv
openai
